worker: further split linux and non-linux features
We've had a few PR's that are specific to our linux-containerd
implementation: #9017 and #9094

After these PR's folks in the community found that they couldn't compile
Concourse for non-Linux platforms. This commit tries to make it easier
for future contributors to make changes to the containerd-linux runtime
and not accidently break things for non-linux platforms.
