+++ worker/runtime/cni_network.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package runtime
 
 import (
