+++ worker/runtime/file_store.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package runtime
 
 import (
