+++ worker/runtime/spec/devices.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package spec
 
 import (
