+++ worker/runtime/spec/mounts.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package spec
 
 import (
