+++ worker/runtime/spec/seccomp.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package spec
 
 import "github.com/opencontainers/runtime-spec/specs-go"
