+++ worker/runtime/spec/spec_nonlinux.go
@@ -1,0 +1,1 @@
+package spec
